requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
packaging==24.0
pillow==10.3.0
PyJWT==2.8.0
sqlparse==0.5.0
whitenoise==6.6.0